ci: run some E2E AWS CI tests with the master using TLS
In order to get CI testing of TLS support, this change enables TLS for
certain tests in the CircleCI config.

To support this, `det-deploy aws up` can now automatically generate and
use a self-signed certificate for the master; it also exposes port 8443
on the instance firewall in addition to port 8080. This ability or
something similar might be good to expose at some point, but for now we
keep it only for internal test usage.

The CircleCI config is updated in a number of places to allow toggling
the TLS support per test and threading that and the follow-up
information (the port and scheme) through to everywhere it's needed.

The remainder of the change is tweaking some of the Python cert handling
code to better handle unverified connections, which is necessary since
the cert is being generated and served without being distributed to the
clients beforehand.